requirements.txt
Pillow==10.4.0
numpy==2.1.3
tabulate==0.9.0
